Using long and IntPtr instead of int for addresses. Remove compare from ISearch. Add virtualQuery for fast searches.

+++ MemoryManager/AddressLock.cs
@@ -7,10 +7,10 @@ namespace MemoryManager
 {
     internal struct AddressLock
     {
-        public int Address;
+        public IntPtr Address;
         public byte[] Data;
 
-        public AddressLock(int address, byte[] data)
+        public AddressLock(IntPtr address, byte[] data)
         {
             Address = address;
             Data = data;
